Restructure sampling for Z
